remove helper text
