--- packages/atlas/src/components/HelperText/HelperText.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, Story } from '@storybook/react'
-import React from 'react'
-
-import { HelperText, HelperTextProps } from './HelperText'
-
-export default {
-  title: 'inputs/HelperText',
-  component: HelperText,
-  argTypes: {
-    helperText: { defaultValue: 'some helper text' },
-  },
-} as Meta
-
-const Template: Story<HelperTextProps> = (args) => <HelperText {...args} />
-
-export const Default = Template.bind({})
