--- packages/atlas/src/components/HelperText/index.ts
@@ -1,1 +1,0 @@
-export * from './HelperText'
